pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sec-api-automation"
version = "0.1.0"
description = "SEC filing downloader and Google Drive uploader"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "tenacity>=8.2.3",
  "beautifulsoup4>=4.12.3",
  "reportlab>=4.2.2",
  "google-api-python-client>=2.151.0",
  "google-auth>=2.36.0",
  "google-auth-oauthlib>=1.2.0",
]

[project.scripts]
sec-api-automation = "sec_api_automation.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
